% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ram_operations.R
\name{associate_resource_share}
\alias{associate_resource_share}
\title{Associates the specified resource share with the specified principals and resources}
\usage{
associate_resource_share(resourceShareArn, resourceArns = NULL,
  principals = NULL, clientToken = NULL)
}
\arguments{
\item{resourceShareArn}{[required] The Amazon Resource Name (ARN) of the resource share.}

\item{resourceArns}{The Amazon Resource Names (ARN) of the resources.}

\item{principals}{The principals.}

\item{clientToken}{A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.}
}
\description{
Associates the specified resource share with the specified principals and resources.
}
\section{Accepted Parameters}{
\preformatted{associate_resource_share(
  resourceShareArn = "string",
  resourceArns = list(
    "string"
  ),
  principals = list(
    "string"
  ),
  clientToken = "string"
)
}
}

