% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ram_operations.R
\name{create_resource_share}
\alias{create_resource_share}
\title{Creates a resource share}
\usage{
create_resource_share(name, resourceArns = NULL, principals = NULL,
  tags = NULL, allowExternalPrincipals = NULL, clientToken = NULL)
}
\arguments{
\item{name}{[required] The name of the resource share.}

\item{resourceArns}{The Amazon Resource Names (ARN) of the resources to associate with the resource share.}

\item{principals}{The principals to associate with the resource share. The possible values are IDs of AWS accounts, the ARN of an OU or organization from AWS Organizations.}

\item{tags}{One or more tags.}

\item{allowExternalPrincipals}{Indicates whether principals outside your organization can be associated with a resource share.}

\item{clientToken}{A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.}
}
\description{
Creates a resource share.
}
\section{Accepted Parameters}{
\preformatted{create_resource_share(
  name = "string",
  resourceArns = list(
    "string"
  ),
  principals = list(
    "string"
  ),
  tags = list(
    list(
      key = "string",
      value = "string"
    )
  ),
  allowExternalPrincipals = TRUE|FALSE,
  clientToken = "string"
)
}
}

