% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ram_operations.R
\name{delete_resource_share}
\alias{delete_resource_share}
\title{Deletes the specified resource share}
\usage{
delete_resource_share(
  resourceShareArn = "string",
  clientToken = "string"
)
}
\arguments{
\item{resourceShareArn}{[required] The Amazon Resource Name (ARN) of the resource share.}

\item{clientToken}{A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.}
}
\description{
Deletes the specified resource share.
}
\examples{

}
