% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ram_operations.R
\name{disassociate_resource_share}
\alias{disassociate_resource_share}
\title{Disassociates the specified principals or resources from the specified resource share}
\usage{
disassociate_resource_share(resourceShareArn, resourceArns = NULL,
  principals = NULL, clientToken = NULL)
}
\arguments{
\item{resourceShareArn}{[required] The Amazon Resource Name (ARN) of the resource share.}

\item{resourceArns}{The Amazon Resource Names (ARN) of the resources.}

\item{principals}{The principals.}

\item{clientToken}{A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.}
}
\description{
Disassociates the specified principals or resources from the specified resource share.
}
\section{Accepted Parameters}{
\preformatted{disassociate_resource_share(
  resourceShareArn = "string",
  resourceArns = list(
    "string"
  ),
  principals = list(
    "string"
  ),
  clientToken = "string"
)
}
}

