% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ram_operations.R
\name{get_resource_share_associations}
\alias{get_resource_share_associations}
\title{Gets the associations for the specified resource share}
\usage{
get_resource_share_associations(associationType,
  resourceShareArns = NULL, resourceArn = NULL, principal = NULL,
  associationStatus = NULL, nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{associationType}{[required] The association type.}

\item{resourceShareArns}{The Amazon Resource Names (ARN) of the resource shares.}

\item{resourceArn}{The Amazon Resource Name (ARN) of the resource.}

\item{principal}{The principal.}

\item{associationStatus}{The status of the association.}

\item{nextToken}{The token for the next page of results.}

\item{maxResults}{The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned \code{nextToken} value.}
}
\description{
Gets the associations for the specified resource share.
}
\section{Accepted Parameters}{
\preformatted{get_resource_share_associations(
  associationType = "PRINCIPAL"|"RESOURCE",
  resourceShareArns = list(
    "string"
  ),
  resourceArn = "string",
  principal = "string",
  associationStatus = "ASSOCIATING"|"ASSOCIATED"|"FAILED"|"DISASSOCIATING"|"DISASSOCIATED",
  nextToken = "string",
  maxResults = 123
)
}
}

