% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ram_operations.R
\name{get_resource_shares}
\alias{get_resource_shares}
\title{Gets the specified resource shares or all of your resource shares}
\usage{
get_resource_shares(
  resourceShareArns = list(
    "string"
  ),
  resourceShareStatus = "PENDING"|"ACTIVE"|"FAILED"|"DELETING"|"DELETED",
  resourceOwner = "SELF"|"OTHER-ACCOUNTS",
  name = "string",
  tagFilters = list(
    list(
      tagKey = "string",
      tagValues = list(
        "string"
      )
    )
  ),
  nextToken = "string",
  maxResults = 123
)
}
\arguments{
\item{resourceShareArns}{The Amazon Resource Names (ARN) of the resource shares.}

\item{resourceShareStatus}{The status of the resource share.}

\item{resourceOwner}{[required] The type of owner.}

\item{name}{The name of the resource share.}

\item{tagFilters}{One or more tag filters.}

\item{nextToken}{The token for the next page of results.}

\item{maxResults}{The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned \code{nextToken} value.}
}
\description{
Gets the specified resource shares or all of your resource shares.
}
\examples{

}
