% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ram_operations.R
\name{list_resources}
\alias{list_resources}
\title{Lists the resources that the specified principal can access}
\usage{
list_resources(
  resourceOwner = "SELF"|"OTHER-ACCOUNTS",
  principal = "string",
  resourceType = "string",
  resourceArns = list(
    "string"
  ),
  resourceShareArns = list(
    "string"
  ),
  nextToken = "string",
  maxResults = 123
)
}
\arguments{
\item{resourceOwner}{[required] The type of owner.}

\item{principal}{The principal.}

\item{resourceType}{The resource type.}

\item{resourceArns}{The Amazon Resource Names (ARN) of the resources.}

\item{resourceShareArns}{The Amazon Resource Names (ARN) of the resource shares.}

\item{nextToken}{The token for the next page of results.}

\item{maxResults}{The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned \code{nextToken} value.}
}
\description{
Lists the resources that the specified principal can access.
}
\examples{

}
