% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ram_package.R
\docType{package}
\name{paws.ram-package}
\alias{paws.ram}
\alias{paws.ram-package}
\title{paws.ram: AWS Resource Access Manager}
\description{
Use AWS Resource Access Manager to share AWS
    resources between AWS accounts.  To share a resource, you create a
    resource share, associate the resource with the resource share, and
    specify the principals that can access the resource.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
