% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ram_operations.R
\name{tag_resource}
\alias{tag_resource}
\title{Adds the specified tags to the specified resource share}
\usage{
tag_resource(resourceShareArn, tags)
}
\arguments{
\item{resourceShareArn}{[required] The Amazon Resource Name (ARN) of the resource share.}

\item{tags}{[required] One or more tags.}
}
\description{
Adds the specified tags to the specified resource share.
}
\section{Accepted Parameters}{
\preformatted{tag_resource(
  resourceShareArn = "string",
  tags = list(
    list(
      key = "string",
      value = "string"
    )
  )
)
}
}

