% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ram_operations.R
\name{untag_resource}
\alias{untag_resource}
\title{Removes the specified tags from the specified resource share}
\usage{
untag_resource(resourceShareArn, tagKeys)
}
\arguments{
\item{resourceShareArn}{[required] The Amazon Resource Name (ARN) of the resource share.}

\item{tagKeys}{[required] The tag keys of the tags to remove.}
}
\description{
Removes the specified tags from the specified resource share.
}
\section{Accepted Parameters}{
\preformatted{untag_resource(
  resourceShareArn = "string",
  tagKeys = list(
    "string"
  )
)
}
}

